feat: Add pipeline verification example and enforce bashrs quality (Refs #105)
New Features:
- Add pipeline_verification.rs example demonstrating verify module
- Add pipeline-verification.md book chapter with full documentation
- Add create-test-apr.md book chapter for utility example

Quality Enforcement:
- Add .bashrsignore with documented rule justifications
- Update CLAUDE.md: bashrs replaces shellcheck for shell quality
- Fix clippy warnings in apr-cli commands (complexity, dead_code)
- Fix Makefile error handling for cargo install commands
- Fix shell scripts for bashrs compliance (quoting, find vs ls)

Dependencies:
- Update renacer 0.7 -> 0.8

+++ crates/apr-cli/src/commands/flow.rs
@@ -333,6 +333,7 @@ fn print_decoder_block(tensor_names: &[String], _verbose: bool) {
 }
 
 /// Print cross-attention flow (critical for debugging Posterior Collapse)
+#[allow(clippy::too_many_lines)] // Visual debugging output requires detailed formatting
 fn print_cross_attention_flow(
     reader: &AprReader,
     tensor_names: &[String],
@@ -625,16 +626,16 @@ fn compute_stats(data: &[f32]) -> (f32, f32, f32, f32) {
 
     let mean = (sum / data.len() as f64) as f32;
 
-    let variance: f64 = data
+    let variance: f32 = (data
         .iter()
         .map(|&x| {
             let diff = f64::from(x) - f64::from(mean);
             diff * diff
         })
         .sum::<f64>()
-        / data.len() as f64;
+        / data.len() as f64) as f32;
 
-    let std = (variance as f32).sqrt();
+    let std = variance.sqrt();
 
     (min, max, mean, std)
 }
